Lines Description and Files Description Done
Every file is described in a paragraph or less, and the important code lines are explained. Note that for the 5 measurement files, since they all share the same mold, if you would like to get a better understanding, read the comments from all 5 files; the shared comments are split between the 5 files.